Jquery - Attempt - Not Successfull
I managed to create something that makes sense to me, in JS file.

JS should fetch on submit's click the input field content, store in a variable, create a url in GET...
Send the info to the server.
Then wait for the reply, and display the reply's content in the dialogbox area.

However this doesn't works.

Direct input in the nav browser, works.
So server works.

It seems that I don't send what's requiered. As after clicking submit, the adress bar in the browser returns:
http://127.0.0.1:5000/?user_request=pouloulouloulou&submit=

instead of : http://127.0.0.1:5000/msg/pouloulou%20du%20pouloulou