Gmap Works! - Static Gmap
I am Using Maps Static API
This API Key needs to be activated.
It returns a png image, with a marker.

ISSUE: If the request isn't found by the Gmap API, it just return an empty image, warning error not found.

This creates also Glitches, when places/cities are named after a person:
e.g. Charlemagne
Wiki returns a biography.
Gmap returns the city of Charlemagne in Quebec.